updated liquibase

+++ src/main/java/com/charles/knightonlineapi/model/dto/UserHeroBasicDTO.java
@@ -15,7 +15,7 @@ public class UserHeroBasicDTO implements Serializable {
 
     private Long id;
     private Integer level;
-    private Integer experience;
+    private Long experience;
     private Integer enhanced;
     private HeroBasicDTO hero;
 }
